apps/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "SynTag API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "3.9.1"
aiohttp-retry = "2.8.3"
aiosignal = "1.3.1"
alembic = "1.12.1"
annotated-types = "0.6.0"
anyio = "3.7.1"
asgiref = "3.7.2"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
attrs = "23.1.0"
blinker = "1.7.0"
certifi = "2023.7.22"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
clickclick = "20.10.2"
connexion = "2.14.2"
cryptography = "42.0.2"
distlib = "0.3.8"
distro = "1.9.0"
ecdsa = "0.18.0"
exceptiongroup = "1.1.3"
fastapi = "0.109.2"
filelock = "3.13.1"
frozenlist = "1.4.1"
greenlet = "3.0.1"
h11 = "0.14.0"
httpcore = "1.0.1"
httptools = "0.6.1"
httpx = "0.25.1"
identify = "2.5.33"
idna = "3.4"
importlib-metadata = "7.0.1"
inflect = "7.0.0"
inflection = "0.5.1"
iniconfig = "2.0.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
jsonschema = "4.19.2"
jsonschema-specifications = "2023.7.1"
mako = "1.2.4"
markupsafe = "2.1.3"
mirakuru = "2.5.2"
multidict = "6.0.4"
nodeenv = "1.8.0"
openai = "^1.36.0"
packaging = "23.2"
pluggy = "1.4.0"
port-for = "0.7.2"
postgres = "4.0"
pre-commit = "3.5.0"
psycopg = "^3.1.3"
psycopg-binary = "3.1.18"
psycopg-pool = "3.2.1"
psycopg2-binary = "^2.9.3"
psycopg2-pool = "1.1"
pyasn1 = "0.5.1"
pycparser = "2.21"
pydantic-settings = "2.2.1"
pyjwt = "2.8.0"
pytest = "^7.1.3"
pytest-alembic = "0.10.7"
pytest-mock = "3.14.0"
pytest-postgresql = "6.0.0"
pytest-sqlalchemy-mock = "0.1.5"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pyyaml = "6.0.1"
referencing = "0.30.2"
regex = "2023.10.3"
requests = "2.31.0"
rpds-py = "0.12.0"
rsa = "4.9"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
starlette = "0.36.3"
strenum = "0.4.15"
stripe = "8.8.0"
swagger-ui-bundle = "0.0.9"
tomli = "2.0.1"
tomlkit = "0.12.3"
tqdm = "4.66.1"
tzdata = "2023.3"
uvicorn = "0.27.1"
uvloop = "0.19.0"
virtualenv = "20.25.0"
watchfiles = "0.21.0"
werkzeug = "2.2.3"
yarl = "1.9.4"
black = "^22.10.0"
isort = "^5.10.1"
responses = "^0.25.0"
urllib3 = "1.26.15"
autopep8 = "^2.1.0"
autoflake = "^2.3.1"
opentelemetry-instrumentation-fastapi = "^0.46b0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-exporter-otlp-proto-http = "^1.25.0"
typing-extensions = "^4.12.2"
pydantic-core = "^2.18.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
langchain-core = "^0.2.33"
langchain-community = "^0.2.12"
azure-ai-documentintelligence = "^1.0.0b3"
langchain-experimental = "^0.0.64"
tiktoken = "^0.7"
langchain-openai = "^0.1.22"
langchain-text-splitters = "^0.2.2"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
markdownify = "^0.13.1"
google-auth = "^2.34.0"
pygments = "^2.18.0"
opentelemetry-instrumentation-sqlalchemy = "^0.46b0"
opentelemetry-instrumentation-httpx = "^0.46b0"
opentelemetry-instrumentation-requests = "^0.46b0"
opentelemetry-instrumentation-logging = "^0.46b0"
twilio = "^9.3.2"
apify-client = "1.8.1"
pytz = "^2024.2"
google-cloud-speech = "^2.27.0"
google-cloud-storage = "^2.18.2"

[tool.poetry.group.dev.dependencies]
aiohttp = "3.9.1"
aiohttp-retry = "2.8.3"
aiosignal = "1.3.1"
alembic = "1.12.1"
annotated-types = "0.6.0"
anyio = "3.7.1"
asgiref = "3.7.2"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
attrs = "23.1.0"
blinker = "1.7.0"
certifi = "2023.7.22"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
clickclick = "20.10.2"
connexion = "2.14.2"
cryptography = "42.0.2"
distlib = "0.3.8"
distro = "1.9.0"
ecdsa = "0.18.0"
exceptiongroup = "1.1.3"
fastapi = "0.109.2"
filelock = "3.13.1"
frozenlist = "1.4.1"
greenlet = "3.0.1"
h11 = "0.14.0"
httpcore = "1.0.1"
httptools = "0.6.1"
httpx = "0.25.1"
identify = "2.5.33"
idna = "3.4"
importlib-metadata = "7.0.1"
inflect = "7.0.0"
inflection = "0.5.1"
iniconfig = "2.0.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
jsonschema = "4.19.2"
jsonschema-specifications = "2023.7.1"
mako = "1.2.4"
markupsafe = "2.1.3"
mirakuru = "2.5.2"
multidict = "6.0.4"
nodeenv = "1.8.0"
openai = "^1.3.6"
packaging = "23.2"
pluggy = "1.4.0"
port-for = "0.7.2"
postgres = "4.0"
pre-commit = "3.5.0"
psycopg = "3.1.18"
psycopg-binary = "3.1.18"
psycopg-pool = "3.2.1"
psycopg2-binary = "2.9.9"
psycopg2-pool = "1.1"
pyasn1 = "0.5.1"
pycparser = "2.21"
pydantic-settings = "2.2.1"
pyjwt = "2.8.0"
pytest = "7.4.3"
pytest-alembic = "0.10.7"
pytest-mock = "3.14.0"
pytest-postgresql = "6.0.0"
pytest-sqlalchemy-mock = "0.1.5"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pyyaml = "6.0.1"
referencing = "0.30.2"
regex = "2023.10.3"
requests = "2.31.0"
rpds-py = "0.12.0"
rsa = "4.9"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
starlette = "0.36.3"
strenum = "0.4.15"
stripe = "8.8.0"
swagger-ui-bundle = "0.0.9"
tomli = "2.0.1"
tomlkit = "0.12.3"
tqdm = "4.66.1"
tzdata = "2023.3"
uvicorn = "0.27.1"
uvloop = "0.19.0"
virtualenv = "20.25.0"
watchfiles = "0.21.0"
werkzeug = "2.2.3"
yarl = "1.9.4"
deptry = "^0.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
